+++ web/ia-search.ts
@@ -1,4 +1,4 @@
-import _ from "lodash";
+import * as _ from "lodash";
 
 const ArchiveOrgCors = "https://cors.archive.org";
