+++ web/main.js
@@ -2,6 +2,7 @@ import _ from "lodash";
 
 import {downloadSnapshotUrlFor, searchForSpectrum} from "./ia-search";
 import {Spectrum, SpectrumShift, SymbolShift} from "./spectrum";
+import {Typist} from "./typist";
 
 const $canvas = document.querySelector("canvas");
 const $effectiveMhz = document.querySelector("#effective-mhz");
@@ -45,6 +46,8 @@ if ($loadGameButton) {
     $searchButton.addEventListener("keyup", debouncedSearch);
 }
 
+const spectrum = new Spectrum($canvas, document.querySelector("#warning"));
+
 let _wasPaused = false;
 
 function pause() {
@@ -71,44 +74,6 @@ document.querySelectorAll("dialog").forEach((dialog) => {
         }
     });
 });
-
-const spectrum = new Spectrum($canvas, document.querySelector("#warning"));
-
-class Typist {
-    constructor(text) {
-        this.counter = 0;
-        let index = 0;
-        this.onCycle = new Map();
-        let counter = 10;
-        while (index < text.length) {
-            let code = text.charCodeAt(index++);
-            if (code >= "A".charCodeAt(0) && code <= "Z".charCodeAt(0)) {
-                this.onCycle.set(counter, {code : SpectrumShift, pressed : true});
-                this.onCycle.set(counter + 7, {code : SpectrumShift, pressed : false});
-                counter += 2;
-                code = code + 32;
-            }
-            if (code === "^".charCodeAt(0)) {
-                this.onCycle.set(counter, {code : SymbolShift, pressed : true});
-                this.onCycle.set(counter + 5, {code : SymbolShift, pressed : false});
-                counter += 2;
-                continue;
-            }
-            if (code === "$".charCodeAt(0))
-                code = 13;
-            this.onCycle.set(counter, {code : code, pressed : true});
-            this.onCycle.set(counter + 3, {code : code, pressed : false});
-            counter += 15;
-        }
-    }
-
-    type() {
-        this.counter++;
-        const todo = this.onCycle.get(this.counter);
-        if (todo)
-            spectrum.setKeyState(todo.code, todo.pressed);
-    }
-}
 
 function updateUrl(updateMap) {
     const parsedQuery = new URLSearchParams(window.location.search);
